Scrolling now possible with render manager camera coordinates. Use BaseRenderComponent.FixedPosition = true to exclude an entity from camera transformation.

+++ Source/Kinectitude/Render/RenderService.cs
@@ -31,10 +31,10 @@ namespace Kinectitude.Render
         private readonly BitmapProperties bitmapProperties;
 
         [PluginProperty("Window Width", "", 800)]
-        public double Width { get; set; }
+        public float Width { get; set; }
 
         [PluginProperty("Window Height", "", 600)]
-        public double Height { get; set; }
+        public float Height { get; set; }
 
         public RenderTarget RenderTarget
         {
